Add tests for updating in various ways
Alter test fixtures to include tracked integer, string and datetime datatypes.

Also add a column which can later be used to test belongs to associations but this isn't intended to be working yet.

These changes were necessary as currently updating fails when multiparameter attributes (e.g. datetime columns) are updated. This is due to the parent column not having the column in its table and ActiveRecord relying on this fact to perform the update.