Add current? and current! methods
Parent instances can now either point at a certain point in time or track the current time (and detail).

The current? method determines which of these states the instance is in.

The current! method forces the instance to track the current time and resets the detail to the current detail should it be pointing at a previous one.

The on= method now forces the instance to point at a certain point in time (and still resets the detail to the detail at that point in time should it be pointing at another one).